requirements.txt
selenium==4.16.0
Pillow==12.1.1
boto3==1.42.49